Rename /universal to /common
