+++ examples/jsm/renderers/common/Animation.js
