+++ examples/jsm/renderers/common/Attributes.js
