+++ examples/jsm/renderers/common/Background.js
