+++ examples/jsm/renderers/common/Bindings.js
