+++ examples/jsm/renderers/common/Buffer.js
