+++ examples/jsm/renderers/common/ChainMap.js
