+++ examples/jsm/renderers/common/Geometries.js
