+++ examples/jsm/renderers/common/Info.js
