+++ examples/jsm/renderers/common/Pipeline.js
