+++ examples/jsm/renderers/common/Pipelines.js
