+++ examples/jsm/renderers/common/ProgrammableStage.js
