+++ examples/jsm/renderers/common/RenderObject.js
