+++ examples/jsm/renderers/common/RenderObjects.js
