+++ examples/jsm/renderers/common/RenderPipelines.js
