+++ examples/jsm/renderers/common/RenderTarget.js
