+++ examples/jsm/renderers/common/Renderer.js
