+++ examples/jsm/renderers/common/SampledTexture.js
