+++ examples/jsm/renderers/common/Sampler.js
