+++ examples/jsm/renderers/common/StorageBuffer.js
