+++ examples/jsm/renderers/common/TextureRenderer.js
