+++ examples/jsm/renderers/common/Textures.js
