+++ examples/jsm/renderers/common/Uniform.js
