+++ examples/jsm/renderers/common/nodes/NodeRender.js
