+++ examples/jsm/renderers/common/nodes/NodeSampledTexture.js
