+++ examples/jsm/renderers/common/nodes/NodeSampler.js
