+++ examples/jsm/renderers/common/nodes/NodeUniform.js
