+++ examples/jsm/renderers/webgpu/backends/webgpu/WebGPUBackend.js
@@ -1,7 +1,7 @@
 import { GPUFeatureName, GPUTextureFormat, GPULoadOp, GPUStoreOp, GPUIndexFormat } from './utils/WebGPUConstants.js';
 
 import WebGPUNodeBuilder from './builder/WGSLNodeBuilder.js';
-import Backend from '../../../universal/Backend.js';
+import Backend from '../../../common/Backend.js';
 
 import { Matrix4, Frustum, DepthFormat } from 'three';
