+++ examples/jsm/renderers/webgpu/backends/webgpu/builder/WGSLNodeBuilder.js
@@ -1,16 +1,16 @@
-import UniformsGroup from '../../../../universal/UniformsGroup.js';
+import UniformsGroup from '../../../../common/UniformsGroup.js';
 import {
 	FloatNodeUniform, Vector2NodeUniform, Vector3NodeUniform, Vector4NodeUniform,
 	ColorNodeUniform, Matrix3NodeUniform, Matrix4NodeUniform
-} from '../../../../universal/nodes/NodeUniform.js';
-import NodeSampler from '../../../../universal/nodes/NodeSampler.js';
-import { NodeSampledTexture, NodeSampledCubeTexture } from '../../../../universal/nodes/NodeSampledTexture.js';
+} from '../../../../common/nodes/NodeUniform.js';
+import NodeSampler from '../../../../common/nodes/NodeSampler.js';
+import { NodeSampledTexture, NodeSampledCubeTexture } from '../../../../common/nodes/NodeSampledTexture.js';
 
-import UniformBuffer from '../../../../universal/UniformBuffer.js';
-import StorageBuffer from '../../../../universal/StorageBuffer.js';
-import { getVectorLength, getStrideLength } from '../../../../universal/BufferUtils.js';
+import UniformBuffer from '../../../../common/UniformBuffer.js';
+import StorageBuffer from '../../../../common/StorageBuffer.js';
+import { getVectorLength, getStrideLength } from '../../../../common/BufferUtils.js';
 
-import RenderTarget from '../../../../universal/RenderTarget.js';
+import RenderTarget from '../../../../common/RenderTarget.js';
 
 import { NodeBuilder, CodeNode, NodeMaterial } from '../../../../../nodes/Nodes.js';
