+++ examples/webgpu_depth_texture.html
@@ -30,7 +30,7 @@
 
 			import WebGPU from 'three/addons/capabilities/WebGPU.js';
 			import WebGPURenderer from 'three/addons/renderers/webgpu/WebGPURenderer.js';
-			import TextureRenderer from 'three/addons/renderers/universal/TextureRenderer.js';
+			import TextureRenderer from 'three/addons/renderers/common/TextureRenderer.js';
 
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
