+++ examples/webgpu_rtt.html
@@ -30,7 +30,7 @@
 
 			import WebGPU from 'three/addons/capabilities/WebGPU.js';
 			import WebGPURenderer from 'three/addons/renderers/webgpu/WebGPURenderer.js';
-			import TextureRenderer from 'three/addons/renderers/universal/TextureRenderer.js';
+			import TextureRenderer from 'three/addons/renderers/common/TextureRenderer.js';
 
 			let camera, scene, renderer;
 			const mouse = new THREE.Vector2();
